+++ src/main/java/com/couchbase/client/dcp/message/DcpExpirationMessage.java
@@ -60,7 +60,7 @@ public enum DcpExpirationMessage {
     return MessageUtil.getCas(buffer);
   }
 
-  public static short partition(final ByteBuf buffer) {
+  public static int partition(final ByteBuf buffer) {
     return MessageUtil.getVbucket(buffer);
   }
